refactor: Modularizar admin views. #30
