+++ src/admin/views/Teacher/CreateTeacherView.tsx
@@ -5,8 +5,8 @@ import { FaUserTie, FaSave } from "react-icons/fa";
 import Card from "../../../shared/components/Card/CardComponent";
 import Button from "../../../shared/components/Button/ButtonComponent";
 import Input from "../../../shared/components/Input/InputComponent";
-import styles from "./CreateTeacherView.module.css";
 import { useTeacher } from "../../hooks/useTeacher";
+import styles from "./CreateTeacherView.module.css";
 
 const CreateTeacherView: React.FC = () => {
   const { loading, registerTeacher } = useTeacher();
