+++ src/admin/views/Teacher/ListTeacherView.tsx
@@ -3,14 +3,13 @@ import type React from "react";
 import { motion } from "framer-motion";
 import { FaUserTie, FaTools } from "react-icons/fa";
 // import { FaSearch, FaEdit, FaTrash, FaEnvelope, FaPhone, FaBook} from "react-icons/fa";
-// import type { Teacher } from "../../types/adminTypes";
 // import Card from "../../../shared/components/Card/CardComponent";
 import Button from "../../../shared/components/Button/ButtonComponent";
 // import Input from "../../../shared/components/Input/InputComponent";
 // import Badge from "../../../shared/components/Badge/BadgeComponent";
-import styles from "./ViewTeachersView.module.css";
+import styles from "./ListTeacherView.module.css";
 
-const ViewTeachersView: React.FC = () => {
+const ListTeacherView: React.FC = () => {
   // const [searchTerm, setSearchTerm] = useState("");
 
   // const teachers: Teacher[] = [
@@ -189,4 +188,4 @@ const ViewTeachersView: React.FC = () => {
   );
 };
 
-export default ViewTeachersView;
+export default ListTeacherView;
